typo

+++ trunk/src/mk/yocto/math/fit/least-squares.cxx
@@ -431,15 +431,16 @@ namespace yocto
             if( !lu.build(alpha) )
                 throw exception("Singular Pade Approximant");
             crout<real_t>::solve(alpha, beta);
+            
             for(size_t i=1;i<=p;++i) P[i] = beta[i];
             for(size_t i=1;i<=q;++i) Q[i] = beta[p+i];
             
             for(size_t i=1;i<=N;++i)
             {
                 const real_t xi = X[i];
-                real_t num = 0;
+                real_t num = REAL(0.0);
                 for(size_t j=1;j<=p;++j) num += P[j]*ipower(xi,j-1);
-                real_t den = 1;
+                real_t den = REAL(1.0);
                 for(size_t j=1;j<=q;++j) den += Q[j]*ipower(xi,j);
                 Z[i] = num/den;
             }
